+++ HISTORY.md
@@ -1,6 +1,7 @@
 ## Version 4.x - in development
 
 BUG: Issue #669 - Wrong Gettext class filename (gani)
+BUG: Issue #671 - THtmlArea4 in TActiveRepeater (ctrlaltca)
 BUG: Issue #674 - Fix TValidationSummary when no validator has an ErrorMessage set and js is disabled (ctrlaltca)
 BUG: Issue #675 - TDatePicker position problems with jQuery 3.3 (ctrlaltca)
 BUG: Issue #683 - TRequiredFieldValidator doesn't work with TActiveListBox (gani, ctrlaltca)