+++ HISTORY.md
@@ -5,6 +5,7 @@ BUG: Issue #671 - THtmlArea4 in TActiveRepeater (ctrlaltca)
 BUG: Issue #674 - Fix TValidationSummary when no validator has an ErrorMessage set and js is disabled (ctrlaltca)
 BUG: Issue #675 - TDatePicker position problems with jQuery 3.3 (ctrlaltca)
 BUG: Issue #683 - TRequiredFieldValidator doesn't work with TActiveListBox (gani, ctrlaltca)
+ENH: Added support for igbinary serializator (ctrlaltca)
 
 ## Version 4.0.1 - Apr 1, 2018