+++ HISTORY.md
@@ -6,6 +6,7 @@ BUG: Issue #674 - Fix TValidationSummary when no validator has an ErrorMessage s
 BUG: Issue #675 - TDatePicker position problems with jQuery 3.3 (ctrlaltca)
 BUG: Issue #683 - TRequiredFieldValidator doesn't work with TActiveListBox (gani, ctrlaltca)
 ENH: Added support for igbinary serializator (ctrlaltca)
+ENH: Added TGlobalizationAutoDetect::AvailableLanguages property to set the list of available cultures on the website (ctrlaltca)
 
 ## Version 4.0.1 - Apr 1, 2018